requirements.txt
boto3>=1.18.59
python-dotenv
google-cloud-storage
google-api-python-client